Calculate word frequency From GO

Calculate word frequency From GO

count_word_from_GO(go_id, term = NULL, exclude_words = NULL)

Arguments

go_id
A vector of GO IDs.

term
The corresponding names or description of terms if the input are not GO terms.

exclude_words
The words that should be excluded.

Details

The input can be simply set with a vector of GO id to go_id argument so that the GO names are automatically extracted. Users can also provide a vector of long names/descriptions by term argument.

If the input is GO id, the following words are excluded: c("via", "protein", "factor", "side", "type", "specific"). They are analyzed by simplifyEnrichment:::all_GO_word_count().

Value

A data frame with words and frequencies.

See also

Examples

gm = readRDS(system.file("extdata", "random_GO_BP_sim_mat.rds", package = "simplifyEnrichment"))
go_id = rownames(gm)
head(count_word_from_GO(go_id))

#>                  word freq
#> regulation regulation  184
#> cell             cell   61
#> negative     negative   61
#> positive     positive   60
#> process       process   59
#> signaling   signaling   45