Calculate word frequency

Calculate word frequency

count_words(term,
    exclude_words = NULL, stop_words = stopwords(),
    min_word_length = 1, tokenizer = 'words', transform_case = tolower,
    remove_numbers = TRUE, remove_punctuation = TRUE, custom_transformer = NULL,
    stemming = FALSE, dictionary = NULL)

Arguments

term
A vector of description texts.

exclude_words
The words that should be excluded.

stop_words
The stop words that should be be removed.

min_word_length
Minimum length of the word to be counted.

tokenizer
The tokenizer function, one of the values accepted by tm::termFreq.

transform_case
The function normalizing lettercase of the words.

remove_numbers
Whether to remove numbers.

remove_punctuation
Whether to remove punctuation.

custom_transformer
Custom function that transforms words.

stemming
Whether to only keep the roots of inflected words.

dictionary
A vector of words to be counted (if given all other words will be excluded).

Details

The text preprocessing followings the instructions from http://www.sthda.com/english/wiki/word-cloud-generator-in-r-one-killer-function-to-do-everything-you-need .

Value

A data frame with words and frequencies.

Examples

gm = readRDS(system.file("extdata", "random_GO_BP_sim_mat.rds", package = "simplifyEnrichment"))
go_id = rownames(gm)
go_term = AnnotationDbi::select(GO.db::GO.db, keys = go_id, columns = "TERM")$TERM

#> 'select()' returned 1:1 mapping between keys and columns

count_words(go_term)

#>                                              word freq
#> regulation                             regulation  180
#> process                                   process   63
#> positive                                 positive   62
#> cell                                         cell   59
#> negative                                 negative   57
#> signaling                               signaling   42
#> pathway                                   pathway   38
#> response                                 response   36
#> protein                                   protein   33
#> development                           development   32
#> transport                               transport   27
#> ion                                           ion   22
#> metabolic                               metabolic   21
#> activity                                 activity   20
#> receptor                                 receptor   19
#> cellular                                 cellular   18
#> assembly                                 assembly   17
#> biosynthetic                         biosynthetic   17
#> catabolic                               catabolic   17
#> dna                                           dna   17
#> morphogenesis                       morphogenesis   16
#> factor                                     factor   15
#> differentiation                   differentiation   13
#> involved                                 involved   13
#> calcium                                   calcium   11
#> production                             production   11
#> transmembrane                       transmembrane   11
#> complex                                   complex   10
#> apoptotic                               apoptotic    9
#> formation                               formation    9
#> growth                                     growth    9
#> membrane                                 membrane    9
#> migration                               migration    9
#> muscle                                     muscle    9
#> adhesion                                 adhesion    8
#> cardiac                                   cardiac    8
#> homeostasis                           homeostasis    8
#> localization                         localization    8
#> organization                         organization    8
#> transcription                       transcription    8
#> acid                                         acid    7
#> activation                             activation    7
#> cycle                                       cycle    7
#> histone                                   histone    7
#> mitochondrion                       mitochondrion    7
#> via                                           via    7
#> import                                     import    6
#> maintenance                           maintenance    6
#> mitochondrial                       mitochondrial    6
#> plasma                                     plasma    6
#> proliferation                       proliferation    6
#> t                                               t    6
#> endothelial                           endothelial    5
#> establishment                       establishment    5
#> methylation                           methylation    5
#> mrna                                         mrna    5
#> neuron                                     neuron    5
#> processing                             processing    5
#> rna                                           rna    5
#> secretion                               secretion    5
#> stimulus                                 stimulus    5
#> atp                                           atp    4
#> blood                                       blood    4
#> checkpoint                             checkpoint    4
#> chemokine                               chemokine    4
#> chromosome                             chromosome    4
#> damage                                     damage    4
#> endosome                                 endosome    4
#> fatty                                       fatty    4
#> golgi                                       golgi    4
#> immune                                     immune    4
#> interleukinmediated           interleukinmediated    4
#> lipoprotein                           lipoprotein    4
#> meiotic                                   meiotic    4
#> metanephric                           metanephric    4
#> microtubule                           microtubule    4
#> mitotic                                   mitotic    4
#> nuclear                                   nuclear    4
#> oxygen                                     oxygen    4
#> particle                                 particle    4
#> polymerase                             polymerase    4
#> replication                           replication    4
#> smooth                                     smooth    4
#> sodium                                     sodium    4
#> transporter                           transporter    4
#> viral                                       viral    4
#> virus                                       virus    4
#> across                                     across    3
#> actin                                       actin    3
#> adaptive                                 adaptive    3
#> adenylate                               adenylate    3
#> along                                       along    3
#> atrial                                     atrial    3
#> binding                                   binding    3
#> body                                         body    3
#> chemotaxis                             chemotaxis    3
#> cord                                         cord    3
#> deubiquitination                 deubiquitination    3
#> division                                 division    3
#> embryonic                               embryonic    3
#> epidermal                               epidermal    3
#> epithelial                             epithelial    3
#> export                                     export    3
#> filament                                 filament    3
#> gene                                         gene    3
#> glutamate                               glutamate    3
#> host                                         host    3
#> initiation                             initiation    3
#> intracellular                       intracellular    3
#> levels                                     levels    3
#> ligand                                     ligand    3
#> macrophage                             macrophage    3
#> male                                         male    3
#> maturation                             maturation    3
#> mirna                                       mirna    3
#> nerve                                       nerve    3
#> phosphorylation                   phosphorylation    3
#> potassium                               potassium    3
#> precursor                               precursor    3
#> renal                                       renal    3
#> repair                                     repair    3
#> sensory                                   sensory    3
#> signal                                     signal    3
#> spinal                                     spinal    3
#> substrate                               substrate    3
#> system                                     system    3
#> telomere                                 telomere    3
#> tissue                                     tissue    3
#> transduction                         transduction    3
#> transition                             transition    3
#> translation                           translation    3
#> tubule                                     tubule    3
#> absorption                             absorption    2
#> acetylation                           acetylation    2
#> albumin                                   albumin    2
#> alternative                           alternative    2
#> amyloid                                   amyloid    2
#> antigen                                   antigen    2
#> arterial                                 arterial    2
#> autophagy                               autophagy    2
#> axis                                         axis    2
#> axon                                         axon    2
#> b                                               b    2
#> bile                                         bile    2
#> bundle                                     bundle    2
#> cc                                             cc    2
#> cellcell                                 cellcell    2
#> cells                                       cells    2
#> central                                   central    2
#> centrosome                             centrosome    2
#> chaperonemediated               chaperonemediated    2
#> circadian                               circadian    2
#> commitment                             commitment    2
#> communication                       communication    2
#> compact                                   compact    2
#> congression                           congression    2
#> copper                                     copper    2
#> cortex                                     cortex    2
#> cytoplasmic                           cytoplasmic    2
#> cytoskeleton                         cytoskeleton    2
#> death                                       death    2
#> depolarization                     depolarization    2
#> depolymerization                 depolymerization    2
#> detection                               detection    2
#> digestive                               digestive    2
#> disassembly                           disassembly    2
#> dorsal                                     dorsal    2
#> elongation                             elongation    2
#> endoderm                                 endoderm    2
#> endoplasmic                           endoplasmic    2
#> epithelium                             epithelium    2
#> erad                                         erad    2
#> extrinsic                               extrinsic    2
#> fate                                         fate    2
#> fiber                                       fiber    2
#> fibroblast                             fibroblast    2
#> fructose                                 fructose    2
#> gland                                       gland    2
#> glomerular                             glomerular    2
#> guidance                                 guidance    2
#> healing                                   healing    2
#> highdensity                           highdensity    2
#> histamine                               histamine    2
#> hk                                             hk    2
#> hormone                                   hormone    2
#> inflammatory                         inflammatory    2
#> inositol                                 inositol    2
#> interleukin                           interleukin    2
#> iron                                         iron    2
#> junction                                 junction    2
#> kinase                                     kinase    2
#> leukocyte                               leukocyte    2
#> lymphoid                                 lymphoid    2
#> matrix                                     matrix    2
#> mediated                                 mediated    2
#> mesenchymal                           mesenchymal    2
#> mesoderm                                 mesoderm    2
#> metaphase                               metaphase    2
#> microglial                             microglial    2
#> modification                         modification    2
#> modulation                             modulation    2
#> molecules                               molecules    2
#> motif                                       motif    2
#> multicellular                       multicellular    2
#> myocardium                             myocardium    2
#> necrosis                                 necrosis    2
#> nervous                                   nervous    2
#> nucleus                                   nucleus    2
#> organism                                 organism    2
#> oxidative                               oxidative    2
#> perception                             perception    2
#> phosphate                               phosphate    2
#> phospholipid                         phospholipid    2
#> plate                                       plate    2
#> polarity                                 polarity    2
#> polymerization                     polymerization    2
#> postsynaptic                         postsynaptic    2
#> potential                               potential    2
#> preinitiation                       preinitiation    2
#> pressure                                 pressure    2
#> promoter                                 promoter    2
#> proteasomal                           proteasomal    2
#> pulmonary                               pulmonary    2
#> reactive                                 reactive    2
#> recognition                           recognition    2
#> recombination                       recombination    2
#> remodeling                             remodeling    2
#> reticulum                               reticulum    2
#> retrograde                             retrograde    2
#> segregation                           segregation    2
#> silencing                               silencing    2
#> sleep                                       sleep    2
#> sleepwake                               sleepwake    2
#> somatic                                   somatic    2
#> species                                   species    2
#> specification                       specification    2
#> sperm                                       sperm    2
#> stress                                     stress    2
#> synaptic                                 synaptic    2
#> synthase                                 synthase    2
#> taste                                       taste    2
#> telomerase                             telomerase    2
#> telomeric                               telomeric    2
#> thelper                                   thelper    2
#> tract                                       tract    2
#> transmission                         transmission    2
#> trisphosphate                       trisphosphate    2
#> tube                                         tube    2
#> tumor                                       tumor    2
#> type                                         type    2
#> tyrosine                                 tyrosine    2
#> ventricular                           ventricular    2
#> vesicle                                   vesicle    2
#> wound                                       wound    2
#> acclimation                           acclimation    1
#> accumbens                               accumbens    1
#> acetylcholine                       acetylcholine    1
#> acrosome                                 acrosome    1
#> acrylamide                             acrylamide    1
#> action                                     action    1
#> activated                               activated    1
#> acutephase                             acutephase    1
#> acylglycerol                         acylglycerol    1
#> adiponectin                           adiponectin    1
#> adipose                                   adipose    1
#> adrenergic                             adrenergic    1
#> alkaline                                 alkaline    1
#> alphabeta                               alphabeta    1
#> aminoacylation                     aminoacylation    1
#> ammonia                                   ammonia    1
#> ampa                                         ampa    1
#> amyloidbeta                           amyloidbeta    1
#> anchoring                               anchoring    1
#> anisomycin                             anisomycin    1
#> anterior                                 anterior    1
#> anterograde                           anterograde    1
#> antigenic                               antigenic    1
#> antimicrobial                       antimicrobial    1
#> apicalbasal                           apicalbasal    1
#> arginase                                 arginase    1
#> arginyltrna                           arginyltrna    1
#> arm                                           arm    1
#> artery                                     artery    1
#> aspartictype                         aspartictype    1
#> assimilation                         assimilation    1
#> asymmetric                             asymmetric    1
#> atpasecoupled                       atpasecoupled    1
#> attachment                             attachment    1
#> auditory                                 auditory    1
#> autophagic                             autophagic    1
#> av                                             av    1
#> baseexcision                         baseexcision    1
#> benzylpenicillin                 benzylpenicillin    1
#> beta                                         beta    1
#> biotinylation                       biotinylation    1
#> blue                                         blue    1
#> bombesin                                 bombesin    1
#> branch                                     branch    1
#> break                                       break    1
#> brexanolone                           brexanolone    1
#> bud                                           bud    1
#> caaxbox                                   caaxbox    1
#> calciumdependent                 calciumdependent    1
#> cameratype                             cameratype    1
#> canonical                               canonical    1
#> cap                                           cap    1
#> capacitation                         capacitation    1
#> carbon                                     carbon    1
#> cardiolipin                           cardiolipin    1
#> carnitine                               carnitine    1
#> cascade                                   cascade    1
#> catabolite                             catabolite    1
#> catecholamine                       catecholamine    1
#> cdpositive                             cdpositive    1
#> cellabiotic                           cellabiotic    1
#> cellmatrix                             cellmatrix    1
#> centriole                               centriole    1
#> centromere                             centromere    1
#> centromeric                           centromeric    1
#> ceramide                                 ceramide    1
#> cerebellar                             cerebellar    1
#> cerebral                                 cerebral    1
#> ceruleus                                 ceruleus    1
#> chamber                                   chamber    1
#> changes                                   changes    1
#> channel                                   channel    1
#> characteristics                   characteristics    1
#> chemical                                 chemical    1
#> chemokinemediated               chemokinemediated    1
#> choline                                   choline    1
#> chondrocyte                           chondrocyte    1
#> chromatid                               chromatid    1
#> cilium                                     cilium    1
#> circulatory                           circulatory    1
#> citrate                                   citrate    1
#> class                                       class    1
#> clathrindependent               clathrindependent    1
#> clearance                               clearance    1
#> cleavage                                 cleavage    1
#> cluster                                   cluster    1
#> coagulation                           coagulation    1
#> coenzyme                                 coenzyme    1
#> cohesion                                 cohesion    1
#> collecting                             collecting    1
#> colonystimulating               colonystimulating    1
#> complement                             complement    1
#> compound                                 compound    1
#> conduction                             conduction    1
#> connecting                             connecting    1
#> constitutive                         constitutive    1
#> contraction                           contraction    1
#> core                                         core    1
#> cornification                       cornification    1
#> cortisol                                 cortisol    1
#> costimulation                       costimulation    1
#> cotranslational                   cotranslational    1
#> coupled                                   coupled    1
#> coupling                                 coupling    1
#> crosslink                               crosslink    1
#> cterminal                               cterminal    1
#> cyclase                                   cyclase    1
#> cyclaseactivating               cyclaseactivating    1
#> cyclaseinhibiting               cyclaseinhibiting    1
#> cytokinemediated                 cytokinemediated    1
#> cytokinesis                           cytokinesis    1
#> cytosol                                   cytosol    1
#> cytotoxicity                         cytotoxicity    1
#> deacetylation                       deacetylation    1
#> decreased                               decreased    1
#> demethylation                       demethylation    1
#> dendritic                               dendritic    1
#> dentincontaining                 dentincontaining    1
#> deoxyribonucleoside           deoxyribonucleoside    1
#> dephosphorylation               dephosphorylation    1
#> detoxification                     detoxification    1
#> developmental                       developmental    1
#> dgtp                                         dgtp    1
#> dibasic                                   dibasic    1
#> dipeptide                               dipeptide    1
#> dispersal                               dispersal    1
#> distal                                     distal    1
#> diterpenoid                           diterpenoid    1
#> diversification                   diversification    1
#> dloop                                       dloop    1
#> dnadependent                         dnadependent    1
#> dnatemplated                         dnatemplated    1
#> doublestrand                         doublestrand    1
#> doxorubicin                           doxorubicin    1
#> dsrna                                       dsrna    1
#> duct                                         duct    1
#> dump                                         dump    1
#> dynein                                     dynein    1
#> early                                       early    1
#> ectodermal                             ectodermal    1
#> edge                                         edge    1
#> electrical                             electrical    1
#> electron                                 electron    1
#> end                                           end    1
#> endocardium                           endocardium    1
#> endochondral                         endochondral    1
#> endocytosis                           endocytosis    1
#> endopeptidase                       endopeptidase    1
#> engulfment                             engulfment    1
#> envelope                                 envelope    1
#> eosinophil                             eosinophil    1
#> epinephrine                           epinephrine    1
#> er                                             er    1
#> ergic                                       ergic    1
#> erythrocyte                           erythrocyte    1
#> esterification                     esterification    1
#> estrogen                                 estrogen    1
#> ethanol                                   ethanol    1
#> eukaryotic                             eukaryotic    1
#> excretion                               excretion    1
#> exogenous                               exogenous    1
#> expression                             expression    1
#> eye                                           eye    1
#> fad                                           fad    1
#> fasciculation                       fasciculation    1
#> fat                                           fat    1
#> fc                                             fc    1
#> female                                     female    1
#> fes                                           fes    1
#> fibril                                     fibril    1
#> fibronectin                           fibronectin    1
#> fluid                                       fluid    1
#> follicle                                 follicle    1
#> forelimb                                 forelimb    1
#> form                                         form    1
#> g                                               g    1
#> generation                             generation    1
#> genitalia                               genitalia    1
#> genome                                     genome    1
#> glucocorticoid                     glucocorticoid    1
#> gluconeogenesis                   gluconeogenesis    1
#> glucose                                   glucose    1
#> glucosylceramide                 glucosylceramide    1
#> glycine                                   glycine    1
#> glyoxylate                             glyoxylate    1
#> gmi                                           gmi    1
#> gonad                                       gonad    1
#> gonadotropin                         gonadotropin    1
#> granule                                   granule    1
#> granuloma                               granuloma    1
#> granzymemediated                 granzymemediated    1
#> guanine                                   guanine    1
#> ha                                             ha    1
#> hair                                         hair    1
#> heart                                       heart    1
#> heat                                         heat    1
#> hematopoietic                       hematopoietic    1
#> hepatic                                   hepatic    1
#> hepatocyte                             hepatocyte    1
#> heterochromatin                   heterochromatin    1
#> hh                                             hh    1
#> hippo                                       hippo    1
#> homeostatic                           homeostatic    1
#> homologous                             homologous    1
#> homophilic                             homophilic    1
#> hr                                             hr    1
#> hs                                             hs    1
#> hydroxyurea                           hydroxyurea    1
#> hypertrophy                           hypertrophy    1
#> hypotonic                               hypotonic    1
#> ii                                             ii    1
#> iii                                           iii    1
#> immunoglobulin                     immunoglobulin    1
#> impulse                                   impulse    1
#> induced                                   induced    1
#> induction                               induction    1
#> inhibition                             inhibition    1
#> inhibitory                             inhibitory    1
#> inner                                       inner    1
#> insulin                                   insulin    1
#> integrin                                 integrin    1
#> integrity                               integrity    1
#> interferonalpha                   interferonalpha    1
#> interkinetic                         interkinetic    1
#> intermediate                         intermediate    1
#> intermediatedensity           intermediatedensity    1
#> intermembrane                       intermembrane    1
#> interneuron                           interneuron    1
#> interstrand                           interstrand    1
#> intrinsic                               intrinsic    1
#> intron                                     intron    1
#> involuntary                           involuntary    1
#> iodide                                     iodide    1
#> ionizing                                 ionizing    1
#> isotype                                   isotype    1
#> joining                                   joining    1
#> kallikreinkinin                   kallikreinkinin    1
#> kappab                                     kappab    1
#> keratinocyte                         keratinocyte    1
#> ketone                                     ketone    1
#> killing                                   killing    1
#> kinasenfkappab                     kinasenfkappab    1
#> lactate                                   lactate    1
#> lateral                                   lateral    1
#> lead                                         lead    1
#> leading                                   leading    1
#> lhistidine                             lhistidine    1
#> light                                       light    1
#> lineage                                   lineage    1
#> lipopolysaccharide             lipopolysaccharide    1
#> lithium                                   lithium    1
#> locus                                       locus    1
#> loop                                         loop    1
#> lowdensity                             lowdensity    1
#> lung                                         lung    1
#> lymphocyte                             lymphocyte    1
#> lymphotoxin                           lymphotoxin    1
#> lysosome                                 lysosome    1
#> magnesium                               magnesium    1
#> malonate                                 malonate    1
#> mammary                                   mammary    1
#> mass                                         mass    1
#> mast                                         mast    1
#> maternal                                 maternal    1
#> mediator                                 mediator    1
#> mediumchain                           mediumchain    1
#> meiosis                                   meiosis    1
#> member                                     member    1
#> mesangial                               mesangial    1
#> mesenchymalepithelial       mesenchymalepithelial    1
#> mesenchyme                             mesenchyme    1
#> methylglyoxal                       methylglyoxal    1
#> minusenddirected                 minusenddirected    1
#> mirnas                                     mirnas    1
#> monophosphate                       monophosphate    1
#> motor                                       motor    1
#> multivesicular                     multivesicular    1
#> muramyl                                   muramyl    1
#> mutation                                 mutation    1
#> myeloid                                   myeloid    1
#> myocyte                                   myocyte    1
#> myotube                                   myotube    1
#> nacetylglucosamine             nacetylglucosamine    1
#> nail                                         nail    1
#> ncrna                                       ncrna    1
#> nephron                                   nephron    1
#> netrinactivated                   netrinactivated    1
#> network                                   network    1
#> neurofilament                       neurofilament    1
#> neuromast                               neuromast    1
#> neuropeptide                         neuropeptide    1
#> neurotransmitter                 neurotransmitter    1
#> neutrophil                             neutrophil    1
#> nfkappabinducing                 nfkappabinducing    1
#> nglycosylase                         nglycosylase    1
#> niknfkappab                           niknfkappab    1
#> nitric                                     nitric    1
#> nodal                                       nodal    1
#> node                                         node    1
#> nonhomologous                       nonhomologous    1
#> nonrem                                     nonrem    1
#> notch                                       notch    1
#> nterminal                               nterminal    1
#> nucleobase                             nucleobase    1
#> nucleoside                             nucleoside    1
#> nucleosome                             nucleosome    1
#> oacetyltransferase             oacetyltransferase    1
#> odontogenesis                       odontogenesis    1
#> oogenesis                               oogenesis    1
#> open                                         open    1
#> optic                                       optic    1
#> orbitofrontal                       orbitofrontal    1
#> organismal                             organismal    1
#> organs                                     organs    1
#> ossification                         ossification    1
#> osteoblast                             osteoblast    1
#> outer                                       outer    1
#> ovarian                                   ovarian    1
#> ovulation                               ovulation    1
#> oxaloacetate                         oxaloacetate    1
#> oxide                                       oxide    1
#> p                                               p    1
#> pain                                         pain    1
#> palate                                     palate    1
#> pancreatic                             pancreatic    1
#> paranodal                               paranodal    1
#> parietal                                 parietal    1
#> pattern                                   pattern    1
#> peptidase                               peptidase    1
#> peptide                                   peptide    1
#> peptidylaspartic                 peptidylaspartic    1
#> peptidylhistidine               peptidylhistidine    1
#> peroxisome                             peroxisome    1
#> ph                                             ph    1
#> phagocytosis                         phagocytosis    1
#> phase                                       phase    1
#> phosphatase                           phosphatase    1
#> phosphatecontaining           phosphatecontaining    1
#> phosphatidylethanolamine phosphatidylethanolamine    1
#> phosphoadenosine                 phosphoadenosine    1
#> phosphosulfate                     phosphosulfate    1
#> photoreceptor                       photoreceptor    1
#> pigmentation                         pigmentation    1
#> placenta                                 placenta    1
#> planar                                     planar    1
#> polyadenylation                   polyadenylation    1
#> polyadpribosylation           polyadpribosylation    1
#> population                             population    1
#> postembryonic                       postembryonic    1
#> preantral                               preantral    1
#> premirna                                 premirna    1
#> preribosome                           preribosome    1
#> presentation                         presentation    1
#> primary                                   primary    1
#> primitive                               primitive    1
#> projection                             projection    1
#> prolactin                               prolactin    1
#> pronucleus                             pronucleus    1
#> proofreading                         proofreading    1
#> proteasome                             proteasome    1
#> proteincoupled                     proteincoupled    1
#> proteindna                             proteindna    1
#> proton                                     proton    1
#> protontransporting             protontransporting    1
#> proximal                                 proximal    1
#> proximaldistal                     proximaldistal    1
#> pseudouridine                       pseudouridine    1
#> purkinje                                 purkinje    1
#> radiation                               radiation    1
#> receptors                               receptors    1
#> recycling                               recycling    1
#> regionalization                   regionalization    1
#> regulatory                             regulatory    1
#> release                                   release    1
#> replicationindependent     replicationindependent    1
#> replicative                           replicative    1
#> repolarization                     repolarization    1
#> respiration                           respiration    1
#> retina                                     retina    1
#> retrotrapezoid                     retrotrapezoid    1
#> reverse                                   reverse    1
#> rho                                           rho    1
#> rhythm                                     rhythm    1
#> ribbon                                     ribbon    1
#> ribonucleoprotein               ribonucleoprotein    1
#> ribosomal                               ribosomal    1
#> right                                       right    1
#> rnabinding                             rnabinding    1
#> rnadependent                         rnadependent    1
#> s                                               s    1
#> salivary                                 salivary    1
#> secondary                               secondary    1
#> secreting                               secreting    1
#> secretory                               secretory    1
#> self                                         self    1
#> semaphorinplexin                 semaphorinplexin    1
#> senescence                             senescence    1
#> septin                                     septin    1
#> septum                                     septum    1
#> sequestering                         sequestering    1
#> serinetype                             serinetype    1
#> sexual                                     sexual    1
#> shortchain                             shortchain    1
#> sister                                     sister    1
#> site                                         site    1
#> skeletal                                 skeletal    1
#> somatostatin                         somatostatin    1
#> sorting                                   sorting    1
#> space                                       space    1
#> spermidine                             spermidine    1
#> splice                                     splice    1
#> spliceosome                           spliceosome    1
#> splicing                                 splicing    1
#> spreading                               spreading    1
#> srpdependent                         srpdependent    1
#> stability                               stability    1
#> stellate                                 stellate    1
#> stem                                         stem    1
#> stimulatory                           stimulatory    1
#> storeoperated                       storeoperated    1
#> stressinduced                       stressinduced    1
#> substratedependent             substratedependent    1
#> succinate                               succinate    1
#> sumo                                         sumo    1
#> sumoylation                           sumoylation    1
#> superfamily                           superfamily    1
#> supramolecular                     supramolecular    1
#> surface                                   surface    1
#> switching                               switching    1
#> symmetric                               symmetric    1
#> sympathetic                           sympathetic    1
#> synthesis                               synthesis    1
#> systemic                                 systemic    1
#> target                                     target    1
#> targeting                               targeting    1
#> telemethylhistidine           telemethylhistidine    1
#> termination                           termination    1
#> testosterone                         testosterone    1
#> thermogenesis                       thermogenesis    1
#> thrombinactivated               thrombinactivated    1
#> thymus                                     thymus    1
#> thyroid                                   thyroid    1
#> tolerance                               tolerance    1
#> tolllike                                 tolllike    1
#> tooth                                       tooth    1
#> torc                                         torc    1
#> touch                                       touch    1
#> trachea                                   trachea    1
#> transcriptase                       transcriptase    1
#> transfer                                 transfer    1
#> transferase                           transferase    1
#> transforming                         transforming    1
#> transsynaptic                       transsynaptic    1
#> triglyceride                         triglyceride    1
#> trimethylhkspecific           trimethylhkspecific    1
#> triphosphate                         triphosphate    1
#> trna                                         trna    1
#> trnatype                                 trnatype    1
#> trophectodermal                   trophectodermal    1
#> trophoblast                           trophoblast    1
#> tubular                                   tubular    1
#> tubulin                                   tubulin    1
#> ubiquitination                     ubiquitination    1
#> ubiquitindependent             ubiquitindependent    1
#> udpnacetylgalactosamine   udpnacetylgalactosamine    1
#> ultradian                               ultradian    1
#> unfolded                                 unfolded    1
#> unwinding                               unwinding    1
#> uracil                                     uracil    1
#> urate                                       urate    1
#> urea                                         urea    1
#> uvb                                           uvb    1
#> valine                                     valine    1
#> vascular                                 vascular    1
#> vdj                                           vdj    1
#> ventricle                               ventricle    1
#> vessel                                     vessel    1
#> virion                                     virion    1
#> wnt                                           wnt    1
#> wounding                                 wounding    1
#> xenophagy                               xenophagy    1
#> xii                                           xii    1