Apply various clustering methods

Apply various clustering methods

cmp_make_clusters(mat, method = setdiff(all_clustering_methods(), "mclust"),
    verbose = TRUE)

Arguments

mat
The similarity matrix.

method
Which methods to compare. All available methods are in all_clustering_methods. A value of all takes all available methods. By default mclust is excluded because its long runtime.

verbose
Whether to print messages.

Details

The function compares following default clustering methods by default:

kmeans

see cluster_by_kmeans.

dynamicTreeCut

see cluster_by_dynamicTreeCut.

mclust

see cluster_by_mclust. By default it is not included.

apcluster

see cluster_by_apcluster.

hdbscan

see cluster_by_hdbscan.

fast_greedy

see cluster_by_igraph.

leading_eigen

see cluster_by_igraph.

louvain

see cluster_by_igraph.

walktrap

see cluster_by_igraph.

MCL

see cluster_by_MCL.

binary_cut

see binary_cut.

Also the user-defined methods in all_clustering_methods are also compared.

Value

A list of cluster label vectors for different clustering methods.

Examples

# \dontrun{
mat = readRDS(system.file("extdata", "random_GO_BP_sim_mat.rds",
    package = "simplifyEnrichment"))
clt = cmp_make_clusters(mat)

#> Cluster 500 terms by 'binary_cut'... 43 clusters, used 0.856966 secs.
#> Cluster 500 terms by 'kmeans'... 14 clusters, used 3.534226 secs.
#> Cluster 500 terms by 'dynamicTreeCut'... 58 clusters, used 0.145431 secs.
#> Cluster 500 terms by 'apcluster'... 39 clusters, used 0.646719 secs.
#> Cluster 500 terms by 'hdbscan'... 13 clusters, used 0.185468 secs.
#> Cluster 500 terms by 'fast_greedy'... 29 clusters, used 0.117954 secs.
#> Cluster 500 terms by 'leading_eigen'... 30 clusters, used 0.551507 secs.
#> Cluster 500 terms by 'louvain'... 29 clusters, used 0.0563972 secs.
#> Cluster 500 terms by 'walktrap'... 26 clusters, used 0.22858 secs.
#> Cluster 500 terms by 'MCL'... 28 clusters, used 2.122988 secs.

# }